workspace_controller/requirements.txt
pyserial
dxcam
pyqt6